Renamed preCICE parameter file
